move to class to lib folder
